mensageria_azure/requirements.txt
azure-servicebus 7.8.1